refactor manager services
